note_driver: time_t outputs in 64-bit format

+++ drivers/note/notesnap_driver.c
@@ -383,12 +383,12 @@ void notesnap_dump_with_stream(FAR struct lib_outstream_s *stream)
 
       perf_convert(note->count, &time);
       lib_sprintf(stream,
-                  "snapshoot: [%u.%09u] "
+                  "snapshoot: [%" PRIu64 ".%09u] "
 #ifdef CONFIG_SMP
                   "[CPU%d] "
 #endif
                   "[%d] %-16s %#" PRIxPTR "\n",
-                  (unsigned)time.tv_sec,
+                  (uint64_t)time.tv_sec,
                   (unsigned)time.tv_nsec,
 #ifdef CONFIG_SMP
                   note->cpu,
